Add files via upload
I have done some major changes in the way simulation to be done. I wrapped repeating parts of the code up to loops to avoid tedious works required for running all parts of the simulation manually. you need to pay attention to comments that I've added to the code as well.